+++ CHANGES.md
@@ -16,6 +16,7 @@ Add new entries to the top of the 1.x.x-NEXT section.
 ## Versions
 
 ### 1.x.x-NEXT (YYYY-MM-DD)
+* Added a `mw.ext.ParserPower.string` Lua library. It contains utility functions for adding (`escape`) or removing (`unescape`) escape sequences from text.
 * …
 
 ### 1.8.0 (2025-07-21)